chapter1
1-2, 1-3 학습 내용 추가
  - 디자인 패턴 예제
  - 개방 폐쇄 원칙
