+++ 1-2.md
@@ -1,0 +1,15 @@
+### 메소드 추출
+
+리펙토링의 한 방법으로 여러 메소드에서 중복되는 관심사를 별도의 메소드로 분리해 내는 것으로 <strong>리팩토링의 한 방법</strong>  
+ [예제코드](./1장/src/main/java/spring/refactoring/ExtractMethod.java)
+
+### 템플릿 메소드 패턴
+
+슈퍼클래스에서 기본적인 로직을 구현하고, 일부 기능에 대해서 추상 메소드나 오버라이딩이 가능한 protected 메소드 등으로 만든 뒤  
+ 서브클래스에서 이런 메소드를 필요에 맞게 구현, 오버라이딩해서 사용하도록 하는 <strong>디자인 패턴</strong>  
+ [예제코드](./1장/src/main/java/spring/pattern/templatemethod/Beverage.java)
+
+### 팩토리 메소드 패턴
+
+서브클래스에서 어떤 오브젝트를 생성할지 결정하게 하는 <strong>디자인 패턴</strong>  
+ [예제코드](./1장/src/main/java/spring/pattern/factorymethod/ShapeFactory.java)
