+++ 1-3.md
@@ -1,0 +1,21 @@
+### 스프링은 무엇을 위한 프레임워크 인가?
+
+스프링은 객체지향적 설계 원칙과 디자인 패턴에 나타난 장점을 자연스럽게 개발자들이 활용할 수 있게 해주는 프레임워크이다.
+
+### 개방 폐쇄 원칙
+
+"소프트웨어 개체(클래스, 모듈, 함수 등등)는 확장에 대해 열려 있어야 하고, 수정에 대해서는 닫혀 있어야 한다"는 <strong>프로그래밍 원칙</strong>  
+ 특정 개체를 수정할 때 그 모듈을 이용하고 있는 모든 개체를 고쳐야 한다면 '개방 폐쇄 원칙'이 잘 지켜지고 있는지 확인해봐야 할 필요가 있다.
+
+> <strong>확장에 대해 열려 있다?</strong>  
+>  새로운 기능이 추가, 수정될 때 해당 모듈을 가져다 쓰는 다른 코드들에 영향을 주지 않으며 기능을 확장할 수 있어야 한다.  
+>  <strong>변경에 대해 닫혀 있다?</strong>  
+>  기존에 작성되어 있는 코드를 수정하지 않고도 애플리케이션의 동작을 추가하거나 변경할 수 있어야 한다.
+
+### 전략 패턴(스트래티지 패턴, Strategy Pattern)
+
+자신의 기능 맥락(context)에서 필요에 따라서 변경이 필요한 알고리즘을 인터페이스를 통해서 통째로 외부로 분리시키고,  
+이를 구현한 구체적인 알고리즘 클래스를 필요에 따라서 바꿔서 사용할 수 있게 하는 디자인 패턴이다.
+(전략 패턴을 구현하게 되면 개방 폐쇄 원칙을 지킬 수 있게 된다.)
+
+[예제코드](./1장/src/main/java/spring/pattern/strategy/Bird.java)
