+++ 1장/src/main/java/spring/pattern/templatemethod/Beverage.java
@@ -1,7 +1,9 @@
-package bomstart.toby.chapter01.pattern.templatemethod;
+package spring.pattern.templatemethod;
 
 /**
  * 템플릿 메소드 패턴
+ * : 슈퍼클래스에서 기본적인 로직을 구현하고, 일부 기능에 대해서 추상 메소드나 오버라이딩이 가능한 protected 메소드 형태로 만든 뒤
+ *   서브클래스에서 필요에 맞게 구현, 오버라이딩해서 사용하도록 하는 디자인 패턴
  */
 public abstract class Beverage {
     public final void makeBeverage() {
@@ -35,6 +37,6 @@ class MixCoffee extends Beverage {
     }
 
     public void addCondiments() {
-        System.out.println("믹스커피 가루 첨가");
+        System.out.println("커피 가루 첨가");
     }
 }
