--- 토비의 스프링/sbj/chapter01/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'chapter01'
